% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Kongoh-package}
\alias{Kongoh}
\alias{Kongoh-package}
\title{Kongoh: A package for probabilistic genotyping based on a continuous model}
\description{
Interpretation of crime stain profiles based on a continuous model.
}
